[backend] refactoring to extract file indexing functions
